requirements.txt
hugchat
python-telegram-bot